Provide different ways to pass options
The username and the watch mount point can now be passed on the command
line, and they will override values from the config file. If an option
is not present in the config file and not set on the command line, then
the user will be prompted to type it.

Interaction with the user can be disabled with the --no-interactive
flag.